refactor: use provider
