--- frontend/packages/app/src/pages/timesheet/team/weekly-approval/types.ts
@@ -1,6 +1,0 @@
-export type WeeklyApprovalProps = {
-  employee: string;
-  startDate: string;
-  open: boolean;
-  onOpenChange: (open: boolean) => void;
-};
